+++ src/functions/functions.test.ts
@@ -1,29 +1,14 @@
 import { removeSubdomainFrom } from './functions';
 
-describe('removeSubdomainFrom', () => {
+describe('removeSubdomainFrom()', () => {
   it.each([
-    {
-      input: 'https://sub.domain.com',
-      output: 'https://domain.com',
-    },
-    {
-      input: 'https://protocol.autonolas.network',
-      output: 'https://autonolas.network',
-    },
-    {
-      input: 'http://protocol.autonolas.network',
-      output: 'http://autonolas.network',
-    },
-    {
-      input: 'https://autonolas.network',
-      output: 'https://autonolas.network',
-    },
-    {
-      input: 'randomString',
-      output: 'randomString',
-    },
+    { input: 'https://sub.domain.com', output: 'https://domain.com' },
+    { input: 'https://registry.olas.network', output: 'https://olas.network' },
+    { input: 'http://registry.olas.network', output: 'http://olas.network' },
+    { input: 'https://autonolas.network', output: 'https://autonolas.network' },
+    { input: 'randomString', output: 'randomString' },
   ])(
-    'should return the correct url for chainId $chainId',
+    'should return the correct url for input $input',
     async ({ input, output }) => {
       const trimmedUrl = removeSubdomainFrom(input);
       expect(trimmedUrl).toBe(output);