+++ src/functions/functions.ts
@@ -82,9 +82,9 @@ export const getExplorerURL = (chainId = 1) => {
     case 97:
       return `https://bscscan.com`;
     case 137:
-      return `https://mumbai.polygonscan.com/`;
-    case 80001:
       return `https://polygonscan.com`;
+    case 80001:
+      return `https://mumbai.polygonscan.com`;
     case 43114:
     case 43113:
       return `https://cchain.explorer.avax.network`;
@@ -95,7 +95,7 @@ export const getExplorerURL = (chainId = 1) => {
     case 100:
       return `https://gnosisscan.io`;
     case 10200:
-      return `https://blockscout.com/xdai/mainnet`;
+      return `https://gnosis.blockscout.com`;
     default:
       return `https://etherscan.io`;
   }